+++ packages/runtime/src/runtime.ts
@@ -23,6 +23,7 @@ import {
 import { convertToString } from "./lib/parsing";
 import { LogicSigAccount } from "./logicsig";
 import { mockSuggestedParams } from "./mock/tx";
+import { convertKeysToHyphens } from "./parser/parser";
 import {
 	AccountAddress,
 	AccountStoreI,
@@ -39,6 +40,7 @@ import {
 	SSCAttributesM,
 	StackElem,
 	State,
+	TxnReceipt,
 	TxReceipt,
 } from "./types";
 // const nacl = require('algosdk/dist/cjs/src/nacl/naclWrappers');
@@ -791,7 +793,7 @@ export class Runtime {
 	 * @param to to address
 	 * @param amount amount of algo in microalgos
 	 */
-	fundLsig(from: RuntimeAccountI, to: AccountAddress, amount: number): TxReceipt {
+	fundLsig(from: RuntimeAccountI, to: AccountAddress, amount: number): TxnReceipt {
 		const fundParam: types.ExecParams = {
 			type: types.TransactionType.TransferAlgo,
 			sign: types.SignType.SecretKey,
@@ -860,7 +862,7 @@ export class Runtime {
 	executeTx(
 		txnParams: types.ExecParams[] | algosdk.SignedTransaction[],
 		debugStack?: number
-	): TxReceipt[] {
+	): TxnReceipt[] {
 		// TODO: union above and create new type in task below:
 		// https://www.pivotaltracker.com/n/projects/2452320/stories/181295625
 		let signedTransactions: algosdk.SignedTransaction[];
@@ -933,11 +935,15 @@ export class Runtime {
 		this.ctx.budget = MAX_APP_PROGRAM_COST * applCallTxNumber;
 		const txReceipts = this.ctx.processTransactions(signedTransactions, appDefMap, lsigMap);
 
+		let txnReceipt: TxnReceipt[] | any
+		for (const txn of txReceipts) {
+			txReceipts.push(convertKeysToHyphens(txn))
+		}
 		// update store only if all the transactions are passed
 		this.store = this.ctx.state;
 
 		// return transaction receipt(s)
-		return txReceipts;
+		return txnReceipt;
 	}
 
 	/**