+++ packages/web/src/lib/myalgowallet-mode.ts
@@ -18,6 +18,7 @@ import {
 	Sign,
 	SignType,
 	TransactionInGroup,
+	TxnReceipt,
 } from "../types";
 import { algoexplorerAlgod } from "./api";
 import { WAIT_ROUNDS } from "./constants";
@@ -139,7 +140,7 @@ export class MyAlgoWalletSession {
 	async sendAndWait(
 		rawTxns: Uint8Array | Uint8Array[],
 		waitRounds = WAIT_ROUNDS
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+	): Promise<TxnReceipt> {
 		const txInfo = await this.algodClient.sendRawTransaction(rawTxns).do();
 		return await this.waitForConfirmation(txInfo.txId, waitRounds);
 	}
@@ -148,12 +149,13 @@ export class MyAlgoWalletSession {
 	private async waitForConfirmation(
 		txId: string,
 		waitRounds = WAIT_ROUNDS
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+	): Promise<TxnReceipt> {
 		const pendingInfo = await algosdk.waitForConfirmation(this.algodClient, txId, waitRounds);
 		if (pendingInfo["pool-error"]) {
 			throw new Error(`Transaction Pool Error: ${pendingInfo["pool-error"] as string}`);
 		}
-		return pendingInfo as algosdk.modelsv2.PendingTransactionResponse;
+		const txnReceipt = { txID: txId, ...pendingInfo }
+		return txnReceipt as TxnReceipt
 	}
 
 	/**
@@ -162,7 +164,7 @@ export class MyAlgoWalletSession {
 	 */
 	async executeTx(
 		execParams: ExecParams[]
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+	): Promise<TxnReceipt> {
 		let signedTxn: SignedTx[] | undefined;
 		let txns: Transaction[] = [];
 		if (execParams.length > 16) {