+++ packages/web/src/lib/web-mode.ts
@@ -9,7 +9,8 @@ import {
 	SignType,
 	TransactionAndSign,
 	TxParams,
-	SignWithMultisig
+	SignWithMultisig,
+	TxnReceipt
 } from "../types";
 import { WAIT_ROUNDS } from "./constants";
 import { log } from "./logger";
@@ -30,11 +31,12 @@ export class WebMode {
 	/**
 	 * wait for confirmation for transaction using transaction id
 	 * @param txId Transaction id
+	 * @param waitRounds number of rounds to wait for transaction to be confirmed - default is 10
 	 */
 	async waitForConfirmation(
 		txId: string,
-		waitRounds = WAIT_ROUNDS
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+		waitRounds: number = WAIT_ROUNDS
+	): Promise<TxnReceipt> {
 		const response = await this.algoSigner.algod({
 			ledger: this.chainName,
 			path: "/v2/status",
@@ -52,7 +54,8 @@ export class WebMode {
 				pendingInfo[CONFIRMED_ROUND] !== null &&
 				(pendingInfo[CONFIRMED_ROUND] as number) > 0
 			) {
-				return pendingInfo as unknown as algosdk.modelsv2.PendingTransactionResponse;
+				const txnReceipt = { txID: txId, ...pendingInfo }
+				return txnReceipt as TxnReceipt
 			}
 			// TODO: maybe we should use "sleep" instead of pinging a node again?
 			currentRound += 1;
@@ -71,14 +74,14 @@ export class WebMode {
 	 */
 	async sendAndWait(
 		signedTxn: string,
-		waitRounds = WAIT_ROUNDS
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+		waitRounds: number = WAIT_ROUNDS
+	): Promise<TxnReceipt> {
 		const txInfo = await this.algoSigner.send({
 			ledger: this.chainName,
 			tx: signedTxn,
 		});
 		if (txInfo && typeof txInfo.txId === "string") {
-			return await this.waitForConfirmation(txInfo.txId);
+			return await this.waitForConfirmation(txInfo.txId, waitRounds);
 		}
 		throw new Error("Transaction Error");
 	}
@@ -216,7 +219,7 @@ export class WebMode {
 	/* eslint-disable sonarjs/cognitive-complexity */
 	async executeTx(
 		transactions: ExecParams[] | TransactionAndSign[]
-	): Promise<algosdk.modelsv2.PendingTransactionResponse> {
+	): Promise<TxnReceipt> {
 		let signedTxn: any;
 		let txns: Transaction[] = [];
 		if (transactions.length > 16 || transactions.length == 0) {